refactor(checker): split type-predicate checking out of type_node.rs (2000-LOC cap)
