+++ crates/tsz-checker/src/types/mod.rs
@@ -27,6 +27,7 @@ mod type_node_property_names;
 mod type_node_query_members;
 mod type_node_resolution;
 mod type_node_signature;
+pub mod type_node_type_predicate;
 pub(crate) mod unique_symbol_arena;
 pub(crate) mod utilities;
 pub(crate) mod window_global_this_annotation;
